vla/requirements-base.txt
# SPDX-License-Identifier: MIT
# SPDX-FileCopyrightText: 2025 nop
#
# Base requirements shared by all VLA backends
# Core PyTorch/JAX are installed separately in Dockerfile

# ==============================================================================
# Transformers stack
# ==============================================================================
transformers==4.57.3
tokenizers==0.22.1
timm==0.9.16
sentencepiece==0.2.1

# ==============================================================================
# Training and fine-tuning
# ==============================================================================
peft>=0.13.0
accelerate>=1.0.0
bitsandbytes>=0.44.0

# ==============================================================================
# Data processing
# ==============================================================================
tensorflow==2.20.0
tensorflow-datasets==4.9.9
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ==============================================================================
# Utilities
# ==============================================================================
einops>=0.7.0
protobuf
rich
draccus==0.10.0
json-numpy
jsonlines>=4.0.0,<5.0.0
PyYAML>=6.0
safetensors>=0.4.0

# ==============================================================================
# Monitoring
# ==============================================================================
wandb>=0.16.0
tqdm>=4.66.0

# ==============================================================================
# LeRobot for OpenPI data format
# ==============================================================================
lerobot>=0.4.0
